+++ Source/JuceBased/ArenaHelpers/Xml/ChaserXmlParser.h
@@ -29,7 +29,7 @@ public:
 	
 	static bool canThisAppVersionLoadThisChaser ( String version, File chaserFile );
     
-    static std::map<int, std::pair<String, int>> getChaserSequences ( File chaserFile, bool onlyActiveChasers );
+    static std::map<int, std::pair<String, int64>> getChaserSequences ( File chaserFile, bool onlyActiveChasers );
 
 private:
 	static XmlElement* getRoot( File chaserFile );
